requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.8
pillow==12.0.0
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
